LocalLog Implementation
Abstracted the log from the logging units. Now the logging units
can consume multiple LocalLogs with different address spaces.
