+++ src/test/java/org/corfudb/infrastructure/LogUnitServerTest.java
@@ -1,6 +1,7 @@
 package org.corfudb.infrastructure;
 
 import com.github.benmanes.caffeine.cache.LoadingCache;
+import org.corfudb.infrastructure.log.LogUnitEntry;
 import org.corfudb.protocols.wireprotocol.LogUnitWriteMsg;
 import org.corfudb.runtime.CorfuRuntime;
 import org.junit.Test;
@@ -36,7 +37,7 @@ public class LogUnitServerTest extends AbstractServerTest {
         m.setPayload(payload);
         sendMessage(m);
 
-        LoadingCache<Long, LogUnitServer.LogUnitEntry> dataCache = s1.getDataCache();
+        LoadingCache<Long, LogUnitEntry> dataCache = s1.getDataCache();
         // Make sure that extra bytes are truncated from the payload byte buf
         assertThat(dataCache.get(address).getBuffer().capacity()).isEqualTo(payload.length);
